Allow user specified top and left positions
This patch allows the user to specify top and left positions
for the ColorBox, rather than using the default behaviour of
centering the box.

Top and left positions are specified using the positionTop
and positionLeft options. They are null by default and any
integer value should work as position. Negative values are
allowed and positions the box outside the viewport.

Signed-off-by: Martin Svangren <martin@masv.net>