use delegate when available
the issue with live() is that it still selects all the elements only to
then discard the selection afterwards. This a) wastes resources and b)
kills IE<8 on really big pages as it doesn't have getElementsByClassName
which causes jQuery to select all elements on check for the class if you
do a $(.classname) selection.

$(document).delegate(...) only selects document which is as fast as you
can possibly be.