Bugfix: Large content and small browser window
When a link at the very end of a website opens colorbox and the content
is very big (high) plus the browser window was resized to a small height before,
the image does not fit in the browser window anymore and the browser window
increases it's document height.
Colorbox now starts to "jump" down vertically out of the visible area of your
browser window everytime when scrolling down to the very end of the webpage and
clicking on the colorbox to switch to the new content-page.